Split up xcbgen-rs/src/defs/mod.rs
This split up the module into multiple files. "pub use" is used so that
the externally visible module structure does not change.

I am not sure that I split this up in a sensible way. Dear reader, feel
free to move things around if you have a better idea.
