+++ xcbgen-rs/src/defs/mod.rs
@@ -6,11 +6,22 @@
 
 use std::cell::RefCell;
 use std::collections::hash_map::Entry as HashMapEntry;
-use std::convert::TryFrom;
 use std::rc::{Rc, Weak};
 
 use fxhash::FxHashMap;
 use once_cell::unsync::OnceCell;
+
+mod alignment;
+mod doc;
+mod expression;
+mod fields;
+mod top_level;
+
+pub use alignment::*;
+pub use doc::*;
+pub use expression::*;
+pub use fields::*;
+pub use top_level::*;
 
 /// A `Module` contains a collection of namespaces.
 #[derive(Debug)]
@@ -346,512 +357,6 @@ impl Import {
     }
 }
 
-/// Any kind of definition in a namespace.
-#[derive(Debug)]
-pub enum Def {
-    /// A request definition.
-    Request(Rc<RequestDef>),
-
-    /// An event definition.
-    Event(EventDef),
-
-    /// An error definition.
-    Error(ErrorDef),
-
-    /// A type definition.
-    Type(TypeDef),
-}
-
-/// A `<request>` definition.
-#[derive(Debug)]
-pub struct RequestDef {
-    /// Reference to the namespace that this request is defined in.
-    pub namespace: Weak<Namespace>,
-
-    /// The name of the request.
-    pub name: String,
-
-    /// The opcode of the request.
-    pub opcode: u8,
-
-    /// Whether adjacent requests of this kind can be combined.
-    ///
-    /// For example, xproto's `PolyRectangle` request contains a list of rectangles to draw. If two
-    /// such requests are sent one after the other with the same fields, their list of rectangles
-    /// can be combined into a single `PolyRectangle` request.
-    pub combine_adjacent: bool,
-
-    /// The value of this request's `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// The list of fields of this request.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-
-    /// The definition of the reply to this request, if any.
-    pub reply: Option<Rc<ReplyDef>>,
-
-    /// The documentation of this request.
-    pub doc: Option<Doc>,
-}
-
-/// A `<reply>` definition.
-#[derive(Debug)]
-pub struct ReplyDef {
-    /// The request that causes this reply.
-    ///
-    /// This reference should always be set after parsing so that `.get()` should not return
-    /// `None`.
-    pub request: OnceCell<Weak<RequestDef>>,
-
-    /// The value of this request's `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// The list of fields of this reply.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-
-    /// The documentation of this reply.
-    pub doc: Option<Doc>,
-}
-
-/// An `<event>` or `<eventcopy>` definition.
-#[derive(Clone, Debug)]
-pub enum EventDef {
-    /// This is an `<event>`.
-    Full(Rc<EventFullDef>),
-
-    /// This is an `<eventcopy>`.
-    Copy(Rc<EventCopyDef>),
-}
-
-impl EventDef {
-    /// The value of the `name` attribute.
-    pub fn name(&self) -> &str {
-        match self {
-            Self::Full(event_full_def) => &event_full_def.name,
-            Self::Copy(event_copy_def) => &event_copy_def.name,
-        }
-    }
-
-    /// The namespace that this event belongs to.
-    pub fn namespace(&self) -> Rc<Namespace> {
-        match self {
-            Self::Full(event_full_def) => event_full_def.namespace.upgrade().unwrap(),
-            Self::Copy(event_copy_def) => event_copy_def.namespace.upgrade().unwrap(),
-        }
-    }
-
-    /// Get the full definition of this event.
-    ///
-    /// For an `<event>`, this simply returns the contained event. For an `<eventcopy>`, this
-    /// returns the definition of the original `<event>` that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if an `<eventcopy>` was not yet resolved.
-    pub fn get_original_full_def(&self) -> Rc<EventFullDef> {
-        match self {
-            Self::Full(event_full_def) => event_full_def.clone(),
-            Self::Copy(event_copy_def) => event_copy_def.get_original_full_def(),
-        }
-    }
-
-    /// Is this event XGE?
-    pub fn is_xge(&self) -> bool {
-        self.get_original_full_def().xge
-    }
-
-    /// Downgrade this definition to a weak reference.
-    ///
-    /// See [`Rc::downgrade`].
-    pub fn as_event_ref(&self) -> EventRef {
-        match self {
-            Self::Full(event_full_def) => EventRef::Full(Rc::downgrade(event_full_def)),
-            Self::Copy(event_copy_def) => EventRef::Copy(Rc::downgrade(event_copy_def)),
-        }
-    }
-}
-
-/// An `<event>` definition.
-#[derive(Debug)]
-pub struct EventFullDef {
-    /// The namespace that this event belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the event.
-    pub name: String,
-
-    /// The `number` of the event.
-    pub number: u16,
-
-    /// The `no-sequence-number` attribute of the event.
-    ///
-    /// The only known event without a sequence number is xproto's `KeymapNotify`.
-    pub no_sequence_number: bool,
-
-    /// The `xge` attribute of the `<event>`.
-    pub xge: bool,
-
-    /// The value of this event's `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// The list of fields of this event.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-
-    /// The documentation of this event.
-    pub doc: Option<Doc>,
-}
-
-/// An `<eventcopy>` definition.
-#[derive(Debug)]
-pub struct EventCopyDef {
-    /// The namespace that this event belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the new event.
-    pub name: String,
-
-    /// The `number` of the new event.
-    pub number: u16,
-
-    /// A reference to the event that is copied.
-    pub ref_: NamedEventRef,
-}
-
-impl EventCopyDef {
-    /// Get the event that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the reference was not resolved yet.
-    pub fn get_original_full_def(&self) -> Rc<EventFullDef> {
-        self.ref_.def.get().unwrap().get_original_full_def()
-    }
-}
-
-/// An `<error>` or `<errorcopy>` definition.
-#[derive(Clone, Debug)]
-pub enum ErrorDef {
-    /// This is an `<error>`.
-    Full(Rc<ErrorFullDef>),
-
-    /// This is an `<errorcopy>`.
-    Copy(Rc<ErrorCopyDef>),
-}
-
-impl ErrorDef {
-    /// The value of the `name` attribute.
-    pub fn name(&self) -> &str {
-        match self {
-            Self::Full(error_full_def) => &error_full_def.name,
-            Self::Copy(error_copy_def) => &error_copy_def.name,
-        }
-    }
-
-    /// The namespace that this error belongs to.
-    pub fn namespace(&self) -> Rc<Namespace> {
-        match self {
-            Self::Full(error_full_def) => error_full_def.namespace.upgrade().unwrap(),
-            Self::Copy(error_copy_def) => error_copy_def.namespace.upgrade().unwrap(),
-        }
-    }
-
-    /// Get the full definition of this error.
-    ///
-    /// For an `<error>`, this simply returns the contained error. For an `<errorcopy>`, this
-    /// returns the definition of the original `<error>` that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if an `<errorcopy>` was not yet resolved.
-    pub fn get_original_full_def(&self) -> Rc<ErrorFullDef> {
-        match self {
-            Self::Full(error_full_def) => error_full_def.clone(),
-            Self::Copy(error_copy_def) => error_copy_def.get_original_full_def(),
-        }
-    }
-}
-
-/// An `<error>` definition.
-#[derive(Debug)]
-pub struct ErrorFullDef {
-    /// The namespace that this error belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the error.
-    pub name: String,
-
-    /// The `number` of the error.
-    // Signed because there are some `-1` somewhere.
-    pub number: i16,
-
-    /// The value of this error's `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// The list of fields of this error.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-}
-
-/// An `<errorcopy>` definition.
-#[derive(Debug)]
-pub struct ErrorCopyDef {
-    /// The namespace that this error belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the new error.
-    pub name: String,
-
-    /// The `number` of the new error.
-    pub number: i16,
-
-    /// A reference to the error that is copied.
-    pub ref_: NamedErrorRef,
-}
-
-impl ErrorCopyDef {
-    /// Get the error that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the reference was not resolved yet.
-    pub fn get_original_full_def(&self) -> Rc<ErrorFullDef> {
-        self.ref_.def.get().unwrap().get_original_full_def()
-    }
-}
-
-/// An enumeration of the possible type definitions that a module can contain.
-///
-/// This enumeration represents any kind of child of the `<xcb>` tag with the exception of
-/// `<request>`, `<event>`, `<eventcopy>`, `<error>`, and `<errorcopy>` because these types have
-/// some special meaning in the protocol.
-#[derive(Debug)]
-pub enum TypeDef {
-    /// A `<struct>` definition.
-    Struct(Rc<StructDef>),
-
-    /// A `<union>` definition.
-    Union(Rc<UnionDef>),
-
-    /// A `<eventstruct>` definition.
-    EventStruct(Rc<EventStructDef>),
-
-    /// A `<xidtype>` definition.
-    Xid(Rc<XidTypeDef>),
-
-    /// A `<xidunion>` definition.
-    XidUnion(Rc<XidUnionDef>),
-
-    /// A `<enum>` definition.
-    Enum(Rc<EnumDef>),
-
-    /// A `<typedef>` definition.
-    Alias(Rc<TypeAliasDef>),
-}
-
-/// A `<struct>` definition.
-#[derive(Debug)]
-pub struct StructDef {
-    /// The namespace that this struct belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the struct.
-    pub name: String,
-
-    /// Alignment information about this struct.
-    ///
-    /// This information is not available before the struct was resolved.
-    pub alignment: OnceCell<ComplexAlignment>,
-
-    /// The list of fields of this struct.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-
-    /// The list of `<paramref>`s that appear in descendants of this struct.
-    ///
-    /// This list is empty before the struct was resolved.
-    pub external_params: RefCell<Vec<ExternalParam>>,
-}
-
-impl StructDef {
-    /// Get the size in bytes of this struct on the wire.
-    ///
-    /// This returns `None` if the struct does not have a fixed size.
-    pub fn size(&self) -> Option<u32> {
-        self.fields
-            .borrow()
-            .iter()
-            .try_fold(0, |sz, field| Some(sz + field.size()?))
-    }
-}
-
-/// A `<union>` definition.
-#[derive(Debug)]
-pub struct UnionDef {
-    /// The namespace that this union belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the union.
-    pub name: String,
-
-    /// Alignment information about this union.
-    ///
-    /// This information is not available before the union was resolved.
-    pub alignment: OnceCell<ComplexAlignment>,
-
-    /// The list of fields of this union.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: Vec<FieldDef>,
-}
-
-impl UnionDef {
-    /// Get the size of the union.
-    ///
-    /// This function returns the size of the largest child of the union.
-    pub fn size(&self) -> u32 {
-        self.fields
-            .iter()
-            .map(|field| field.size().unwrap())
-            .max()
-            .unwrap()
-    }
-}
-
-/// A `<eventstruct>` definition.
-#[derive(Clone, Debug)]
-pub struct EventStructDef {
-    /// The namespace that this eventstruct belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The `name` of the eventstruct.
-    pub name: String,
-
-    /// The list of `<allowed>` children.
-    pub alloweds: Vec<EventStructAllowed>,
-}
-
-/// A `<allowed>` definition inside of an `<eventstruct>`.
-#[derive(Clone, Debug)]
-pub struct EventStructAllowed {
-    /// The extension that events come from.
-    pub extension: String,
-
-    /// Whether XGEs are allowed.
-    pub xge: bool,
-
-    /// The minimum opcode of the events.
-    pub opcode_min: u16,
-
-    /// The maximum opcode of the events.
-    pub opcode_max: u16,
-
-    /// The list of events that this `<allowed>` references.
-    ///
-    /// This member is set up during resolution
-    pub resolved: RefCell<Vec<EventRef>>,
-}
-
-/// An `<xidtype>` definition.
-#[derive(Debug)]
-pub struct XidTypeDef {
-    /// The namespace that this XID type belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The name of the new XID type.
-    pub name: String,
-}
-
-/// An `<xidunion>` definition.
-#[derive(Debug)]
-pub struct XidUnionDef {
-    /// The namespace that this XID union belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The name of the new XID type.
-    pub name: String,
-
-    /// The list of types that are contained in the union.
-    pub types: Vec<NamedTypeRef>,
-}
-
-/// An `<enum>` definition.
-#[derive(Debug)]
-pub struct EnumDef {
-    /// The namespace that this XID union belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The name of the new XID type.
-    pub name: String,
-
-    /// The list of items in this enum.
-    pub items: Vec<EnumItem>,
-
-    /// The documentation of this enum.
-    pub doc: Option<Doc>,
-}
-
-/// An `<item>` inside of an `<enum>`.
-#[derive(Debug)]
-pub struct EnumItem {
-    /// The name of this item.
-    pub name: String,
-
-    /// The value of this item.
-    pub value: EnumValue,
-}
-
-/// The value of an enum item
-#[derive(Debug)]
-pub enum EnumValue {
-    /// A `<value>`.
-    Value(u32),
-
-    /// A `<bit>`.
-    ///
-    /// A bit encodes a value of `1 << bit`.
-    Bit(u8),
-}
-
-/// A `<typedef>` definition.
-#[derive(Debug)]
-pub struct TypeAliasDef {
-    /// The namespace that this typedef belongs to.
-    pub namespace: Weak<Namespace>,
-
-    /// The name of the new type.
-    pub new_name: String,
-
-    /// The name of the old type that is given a new name.
-    pub old_name: NamedTypeRef,
-}
-
-impl TypeAliasDef {
-    /// Get the original type that is copied in this type def.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the type definition was not yet resolved.
-    pub fn get_original_type(&self) -> TypeRef {
-        let mut type_ref = self.old_name.def.get().unwrap().clone();
-        loop {
-            match type_ref {
-                TypeRef::Alias(alias) => {
-                    type_ref = alias.upgrade().unwrap().old_name.def.get().unwrap().clone();
-                }
-                _ => return type_ref,
-            }
-        }
-    }
-}
-
 /// Information about a `<paramref>` that is contained in some expression.
 #[derive(Clone, Debug)]
 pub struct ExternalParam {
@@ -860,1429 +365,4 @@ pub struct ExternalParam {
 
     /// The type of the parameter that is referenced.
     pub type_: TypeRef,
-}
-
-/// A reference to an event by name.
-#[derive(Debug)]
-pub struct NamedEventRef {
-    /// The name of the event that is referenced.
-    pub name: String,
-
-    /// The definition of the event that is referenced.
-    ///
-    /// This field is only set up during resolving.
-    pub def: OnceCell<EventRef>,
-}
-
-impl NamedEventRef {
-    /// Create a new unresolved instance.
-    pub fn unresolved(name: String) -> Self {
-        Self {
-            name,
-            def: OnceCell::new(),
-        }
-    }
-}
-
-/// A reference to an event.
-#[derive(Clone, Debug)]
-pub enum EventRef {
-    /// A reference to an `<event>`.
-    Full(Weak<EventFullDef>),
-
-    /// A reference to an `<eventcopy>`.
-    Copy(Weak<EventCopyDef>),
-}
-
-impl EventRef {
-    /// Get the full definition of this event.
-    ///
-    /// For an `<event>`, this simply returns the contained event. For an `<eventcopy>`, this
-    /// returns the definition of the original `<event>` that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if an `<eventcopy>` was not yet resolved.
-    pub fn get_original_full_def(&self) -> Rc<EventFullDef> {
-        match self {
-            Self::Full(event_full_def) => event_full_def.upgrade().unwrap(),
-            Self::Copy(event_copy_def) => {
-                let mut event_copy_def = event_copy_def.upgrade().unwrap();
-                loop {
-                    match event_copy_def.ref_.def.get().unwrap() {
-                        EventRef::Full(event_full_def) => return event_full_def.upgrade().unwrap(),
-                        EventRef::Copy(event_copy_ref) => {
-                            event_copy_def = event_copy_ref.upgrade().unwrap()
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Is this an XGE?
-    pub fn is_xge(&self) -> bool {
-        self.get_original_full_def().xge
-    }
-
-    /// Upgrade this event reference to an event definition.
-    ///
-    /// See [`Weak::Upgrade`] for more about what this really does.
-    pub fn as_event_def(&self) -> EventDef {
-        match self {
-            Self::Full(event_full_def) => EventDef::Full(event_full_def.upgrade().unwrap()),
-            Self::Copy(event_copy_def) => EventDef::Copy(event_copy_def.upgrade().unwrap()),
-        }
-    }
-}
-
-/// A reference to an error by name.
-#[derive(Debug)]
-pub struct NamedErrorRef {
-    /// The name of the error that is referenced.
-    pub name: String,
-
-    /// The definition of the event that is referenced.
-    ///
-    /// This field is only set up during resolving.
-    pub def: OnceCell<ErrorRef>,
-}
-
-impl NamedErrorRef {
-    /// Create a new unresolved instance.
-    pub fn unresolved(name: String) -> Self {
-        Self {
-            name,
-            def: OnceCell::new(),
-        }
-    }
-}
-
-/// A reference to an error.
-#[derive(Clone, Debug)]
-pub enum ErrorRef {
-    /// A reference to an `<error>`.
-    Full(Weak<ErrorFullDef>),
-
-    /// A reference to an `<errorcopy>`.
-    Copy(Weak<ErrorCopyDef>),
-}
-
-impl ErrorRef {
-    /// Get the full definition of this error.
-    ///
-    /// For an `<error>`, this simply returns the contained error. For an `<errorcopy>`, this
-    /// returns the definition of the original `<error>` that was copied.
-    ///
-    /// # Panics
-    ///
-    /// Panics if an `<errorcopy>` was not yet resolved.
-    pub fn get_original_full_def(&self) -> Rc<ErrorFullDef> {
-        match self {
-            Self::Full(error_full_def) => error_full_def.upgrade().unwrap(),
-            Self::Copy(error_copy_def) => {
-                let mut error_copy_def = error_copy_def.upgrade().unwrap();
-                loop {
-                    match error_copy_def.ref_.def.get().unwrap() {
-                        ErrorRef::Full(error_full_def) => return error_full_def.upgrade().unwrap(),
-                        ErrorRef::Copy(error_copy_ref) => {
-                            error_copy_def = error_copy_ref.upgrade().unwrap()
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Upgrade this error reference to an error definition.
-    ///
-    /// See [`Weak::Upgrade`] for more about what this really does.
-    pub fn as_error_def(&self) -> ErrorDef {
-        match self {
-            Self::Full(error_full_def) => ErrorDef::Full(error_full_def.upgrade().unwrap()),
-            Self::Copy(error_copy_def) => ErrorDef::Copy(error_copy_def.upgrade().unwrap()),
-        }
-    }
-}
-
-/// A reference to a type by name.
-#[derive(Debug)]
-pub struct NamedTypeRef {
-    /// The name of the type that is referenced
-    pub name: String,
-
-    /// The definition of the type that is referenced.
-    ///
-    /// This field is only set up during resolving.
-    pub def: OnceCell<TypeRef>,
-}
-
-impl NamedTypeRef {
-    /// Create a new unresolved instance.
-    pub fn unresolved(name: String) -> Self {
-        Self {
-            name,
-            def: OnceCell::new(),
-        }
-    }
-}
-
-/// An alignment specification.
-///
-/// This structure represents a requirement that some byte position `pos` satisfies
-/// `pos % align == offset`.
-///
-/// `align` must be a power of 2 and `offset` must be less than `align`.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct Alignment {
-    align: u32,
-    offset: u32,
-}
-
-impl Alignment {
-    /// Creates a new `Alignment` with `align` and `offset`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `align` is not a power of two or if `offset` is
-    /// equal or greater than `align`.
-    pub fn new(align: u32, offset: u32) -> Self {
-        assert!(align.is_power_of_two() && offset < align);
-        Self { align, offset }
-    }
-
-    /// Returns the value of `align`.
-    #[inline]
-    pub fn align(self) -> u32 {
-        self.align
-    }
-
-    /// Returns the value of `offset`.
-    #[inline]
-    pub fn offset(self) -> u32 {
-        self.offset
-    }
-
-    /// Advance this alignment specification by some variably sized object.
-    ///
-    /// The resulting value describes the alignment at the end of the variably sized object if it
-    /// is aligned by `self` at its beginning.
-    pub fn advance_variable_size(self, size: VariableSize) -> Self {
-        let align = if size.incr == 0 {
-            self.align
-        } else {
-            self.align.min(size.incr)
-        };
-        let offset = self.offset.wrapping_add(size.base) % align;
-        Self { align, offset }
-    }
-
-    /// Returns an alignment that meets `self` and `other`.
-    pub fn union(self, other: Self) -> Option<Self> {
-        match self.align.cmp(&other.align) {
-            std::cmp::Ordering::Less => {
-                if (other.offset % self.align) != self.offset {
-                    None
-                } else {
-                    Some(other)
-                }
-            }
-            std::cmp::Ordering::Equal => {
-                if self.offset != other.offset {
-                    None
-                } else {
-                    Some(self)
-                }
-            }
-            std::cmp::Ordering::Greater => {
-                if (self.offset % other.align) != other.align {
-                    None
-                } else {
-                    Some(self)
-                }
-            }
-        }
-    }
-
-    /// Returns an alignment that is met by `self` and `other`.
-    pub fn intersection(self, other: Self) -> Self {
-        let align = self.align.min(other.align);
-        let offset1 = self.offset % align;
-        let offset2 = other.offset % align;
-
-        if offset1 != offset2 {
-            let align1 = (1u32 << offset1.trailing_zeros().min(31)).min(align);
-            let align2 = (1u32 << offset2.trailing_zeros().min(31)).min(align);
-            Self {
-                align: align1.min(align2),
-                offset: 0,
-            }
-        } else {
-            Self {
-                align,
-                offset: offset1,
-            }
-        }
-    }
-
-    /// Returns whether `self` meets the alignment requirements
-    /// of `required`.
-    pub fn meets(self, required: Self) -> bool {
-        // `self.align >= required.align` is equivalent to
-        // `self.align % required.align == 0` because `align`
-        // is always a power of 2.
-        self.align >= required.align && (self.offset % required.align) == required.offset
-    }
-}
-
-/// Represents the size of an object that has a variable size.
-///
-/// An object has a minimum size that is described by `base`. It can then have a variable number of
-/// increments of size `incr`. That is, the object can have a size of `base + incr * n` for some
-/// non-negative value of `n`.
-///
-/// `incr` must be zero or a power of 2
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct VariableSize {
-    base: u32,
-    incr: u32,
-}
-
-impl VariableSize {
-    /// Creates a new `VariableSize` with `base` and `incr`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `incr` neither zero nor a power of 2.
-    pub fn new(base: u32, incr: u32) -> Self {
-        assert!(incr == 0 || incr.is_power_of_two());
-        Self { base, incr }
-    }
-
-    /// Returns the value of `base`.
-    #[inline]
-    pub fn base(self) -> u32 {
-        self.base
-    }
-
-    /// Returns the value of `incr`.
-    #[inline]
-    pub fn incr(self) -> u32 {
-        self.incr
-    }
-
-    /// Get the minimum of two values, but not zero (unless both are zero).
-    fn incr_union(incr1: u32, incr2: u32) -> u32 {
-        if incr1 == 0 {
-            incr2
-        } else if incr2 == 0 {
-            incr1
-        } else {
-            incr1.min(incr2)
-        }
-    }
-
-    /// Reduce the base by the given increment.
-    fn reduce_base(base: u32, incr: u32) -> u32 {
-        if incr == 0 {
-            base
-        } else {
-            base % incr
-        }
-    }
-
-    /// Create an instance that describes a zero-sized type.
-    #[inline]
-    pub fn zero() -> Self {
-        Self { base: 0, incr: 0 }
-    }
-
-    /// Return a description of the size of things when `self` is appended to `other`.
-    // FIXME: This code makes no sense to me. When I have something of size `1+4*n` and something
-    // of size `2+8*m`, the result has sizes `3+4*n+8*m` and not `3+4*l`. There is a discrepancy
-    // between what the docs for `VariableSize` say what it represents and what this method is
-    // supposed to do.
-    pub fn append(self, other: Self) -> Self {
-        Self {
-            // FIXME: check overflow
-            base: self.base + other.base,
-            incr: Self::incr_union(self.incr, other.incr),
-        }
-    }
-
-    /// Returns an `VariableSize` that can represent all the sizes
-    /// represented by `self` and `other`.
-    pub fn union(self, other: Self) -> Self {
-        let incr_union = Self::incr_union(self.incr, other.incr);
-        if self.base == other.base {
-            Self {
-                base: self.base,
-                incr: incr_union,
-            }
-        } else {
-            let base1 = Self::reduce_base(self.base, incr_union);
-            let base2 = Self::reduce_base(other.base, incr_union);
-            if base1 == base2 {
-                Self {
-                    base: base1,
-                    incr: incr_union,
-                }
-            } else if base1 == 0 || base2 == 0 {
-                // FIXME: I am quite confused by this code. Some quick println!-debugging says that
-                // this is only called with self == Self::zero(). In this case it returns base:0,
-                // incr: other.base.
-                Self {
-                    base: 0,
-                    incr: 1u32 << (base1 | base2).trailing_zeros().min(31),
-                }
-            } else {
-                // FIXME: I am quite confused by this code.
-                let min_base = base1.min(base2);
-                let max_base = base1.max(base2);
-                let incr1 = 1u32 << min_base.trailing_zeros().min(31);
-                let incr2 = 1u32 << (max_base - min_base).trailing_zeros().min(31);
-                if incr1 > incr2 {
-                    Self {
-                        base: max_base - min_base,
-                        incr: incr1,
-                    }
-                } else {
-                    Self {
-                        base: min_base,
-                        incr: incr2,
-                    }
-                }
-                /*Self {
-                    base: 0,
-                    incr: 1u32 << (base1 | base2).trailing_zeros().min(31),
-                }*/
-            }
-        }
-    }
-
-    /// Describe the size of an arbitrary number of elements.
-    pub fn zero_one_or_many(self) -> Self {
-        // Self represents sizes `base + incr * n`, where `n >= 0`.
-        // The returned value must represent sizes `(base + incr * n) * m`
-        // (or a superset), where `m >= 0`.
-        if self.base == 0 && self.incr == 0 {
-            Self::zero()
-        } else {
-            // `(base + incr * n) * m = base * m + incr * n * m = gcd(base, incr) * l`
-            Self {
-                base: 0,
-                // `self.base | self.incr` won't be zero, so `trailing_zeros < 32`.
-                incr: 1u32 << (self.base | self.incr).trailing_zeros(),
-            }
-        }
-    }
-
-    /// Describe the size of `n` elements of this type.
-    // FIXME: I am not quite sure what is going on here, but a quick println!()-debugging says this
-    // is only called with incr:0 and n > 0 and in that case it returns base:base*n and incr:0.
-    pub fn repeat_n(self, n: u32) -> Self {
-        if n == 0 {
-            Self::zero()
-        } else {
-            let (base, base_overflow) = self
-                .base
-                .checked_mul(n)
-                .map(|base| (base, false))
-                .unwrap_or_else(|| Self::reduce_base(self.base, self.incr).overflowing_mul(n));
-
-            let (mut incr, incr_overflow) = self.incr.overflowing_mul(n);
-            if incr_overflow {
-                incr = 1 << 31;
-            }
-
-            if base_overflow {
-                incr = 1 << (incr | base | 1 << 31).trailing_zeros().min(31);
-                Self { base: 0, incr }
-            } else {
-                if incr != 0 {
-                    incr = 1 << incr.trailing_zeros().min(31);
-                }
-                Self { base, incr }
-            }
-        }
-    }
-}
-
-/// Some computed alignment information.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct ComplexAlignment {
-    /// Alignment at the beginning of the structure.
-    pub begin: Alignment,
-    /// Alignment at the end of the structure.
-    pub body: AlignBody,
-    /// Internal alignments made inside the object using
-    /// align pads.
-    // FIXME: What does that mean?
-    pub internal_align: u32,
-}
-
-/// Alignment information after some structure.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum AlignBody {
-    /// Some `VariableSize` can describe the size of the body.
-    Size(VariableSize),
-
-    /// Information about the alignment after the structure.
-    EndAlign(Alignment),
-}
-
-impl ComplexAlignment {
-    /// Create alignment information for some structure with a fixed size `size` that needs to be
-    /// aligned to a multiple of `align`.
-    #[inline]
-    pub fn fixed_size(size: u32, align: u32) -> Self {
-        Self {
-            begin: Alignment { align, offset: 0 },
-            body: AlignBody::Size(VariableSize::new(size, 0)),
-            internal_align: 1,
-        }
-    }
-
-    /// Create alignment information for an empty structure.
-    #[inline]
-    pub fn zero_sized() -> Self {
-        Self::fixed_size(0, 1)
-    }
-
-    /// Get the alignment after the structure.
-    pub fn end_align(self) -> Alignment {
-        match self.body {
-            AlignBody::Size(size) => self.begin.advance_variable_size(size),
-            AlignBody::EndAlign(end_align) => end_align,
-        }
-    }
-
-    /// Compute alignment information for an arbitrary repetition of the current description.
-    ///
-    /// This returns `None` if the alignment of this structure is violated, i.e. it cannot be
-    /// repeated.
-    pub fn zero_one_or_many(self) -> Option<Self> {
-        if !self.end_align().meets(self.begin) {
-            None
-        } else {
-            match self.body {
-                AlignBody::Size(size) => Some(Self {
-                    begin: self.begin,
-                    body: AlignBody::Size(size.zero_one_or_many()),
-                    internal_align: self.internal_align,
-                }),
-                AlignBody::EndAlign(end_align) => Some(Self {
-                    begin: self.begin,
-                    body: AlignBody::EndAlign(self.begin.union(end_align)?),
-                    internal_align: self.internal_align,
-                }),
-            }
-        }
-    }
-
-    /// Compute alignment information for `n` repetitions of this type.
-    ///
-    /// This returns `None` if the alignment of this structure is violated, i.e. it cannot be
-    /// repeated.
-    pub fn repeat_n(self, n: u32) -> Option<Self> {
-        if n == 0 {
-            Some(Self {
-                begin: self.begin,
-                body: AlignBody::Size(VariableSize::zero()),
-                internal_align: self.internal_align,
-            })
-        } else if n > 1 && !self.end_align().meets(self.begin) {
-            None
-        } else {
-            match self.body {
-                AlignBody::Size(size) => Some(Self {
-                    begin: self.begin,
-                    body: AlignBody::Size(size.repeat_n(n)),
-                    internal_align: self.internal_align,
-                }),
-                AlignBody::EndAlign(end_align) => Some(Self {
-                    begin: self.begin,
-                    body: AlignBody::EndAlign(end_align),
-                    internal_align: self.internal_align,
-                }),
-            }
-        }
-    }
-
-    /// Get a new alignment description for the combination of two types right after each other.
-    pub fn append(self, next: Self) -> Option<Self> {
-        if !self.end_align().meets(next.begin) {
-            None
-        } else {
-            let new_body = match (self.body, next.body) {
-                (AlignBody::Size(curr_size), AlignBody::Size(next_size)) => {
-                    AlignBody::Size(curr_size.append(next_size))
-                }
-                (AlignBody::EndAlign(curr_end_align), AlignBody::Size(next_size)) => {
-                    AlignBody::EndAlign(curr_end_align.advance_variable_size(next_size))
-                }
-                (_, AlignBody::EndAlign(next_end_align)) => AlignBody::EndAlign(next_end_align),
-            };
-            Some(Self {
-                begin: self.begin,
-                body: new_body,
-                internal_align: self.internal_align.max(next.internal_align),
-            })
-        }
-    }
-
-    /// Compute a new alignment description for two types appearing in a union, i.e. both starting
-    /// at the same position.
-    pub fn union_append(self, other: Self) -> Option<Self> {
-        if !self.begin.meets(other.begin) {
-            None
-        } else {
-            let new_body = match (self.body, other.body) {
-                (AlignBody::Size(curr_size), AlignBody::Size(next_size)) => {
-                    AlignBody::Size(curr_size.union(next_size))
-                }
-                (AlignBody::Size(curr_size), AlignBody::EndAlign(next_end_align)) => {
-                    AlignBody::EndAlign(
-                        self.begin
-                            .advance_variable_size(curr_size)
-                            .intersection(next_end_align),
-                    )
-                }
-                (AlignBody::EndAlign(curr_end_align), AlignBody::Size(next_size)) => {
-                    AlignBody::EndAlign(
-                        curr_end_align.intersection(self.begin.advance_variable_size(next_size)),
-                    )
-                }
-                (AlignBody::EndAlign(curr_end_align), AlignBody::EndAlign(next_end_align)) => {
-                    AlignBody::EndAlign(curr_end_align.intersection(next_end_align))
-                }
-            };
-            Some(Self {
-                begin: self.begin,
-                body: new_body,
-                internal_align: self.internal_align.max(other.internal_align),
-            })
-        }
-    }
-
-    /// Compute alignment information of types that can appear optionally after each other.
-    ///
-    /// This computes the alignment for the concatenation of bitcases.
-    pub fn bitcase_append(self, next: Self) -> Option<Self> {
-        if !self.end_align().meets(next.begin) {
-            None
-        } else {
-            let new_body = match (self.body, next.body) {
-                (AlignBody::Size(curr_size), AlignBody::Size(next_size)) => {
-                    AlignBody::Size(curr_size.union(curr_size.append(next_size)))
-                }
-                (AlignBody::Size(curr_size), AlignBody::EndAlign(next_end_align)) => {
-                    AlignBody::EndAlign(
-                        self.begin
-                            .advance_variable_size(curr_size)
-                            .intersection(next_end_align),
-                    )
-                }
-                (AlignBody::EndAlign(curr_end_align), AlignBody::Size(next_size)) => {
-                    AlignBody::EndAlign(
-                        curr_end_align
-                            .intersection(curr_end_align.advance_variable_size(next_size)),
-                    )
-                }
-                (AlignBody::EndAlign(curr_end_align), AlignBody::EndAlign(next_end_align)) => {
-                    AlignBody::EndAlign(curr_end_align.intersection(next_end_align))
-                }
-            };
-            Some(Self {
-                begin: self.begin,
-                body: new_body,
-                internal_align: self.internal_align.max(next.internal_align),
-            })
-        }
-    }
-}
-
-/// Some field of a complex type.
-#[derive(Debug)]
-pub enum FieldDef {
-    /// A `<pad>`.
-    Pad(PadField),
-
-    /// A `<field>`.
-    Normal(NormalField),
-
-    /// A `<list>` that is not a list of file descripts.
-    List(ListField),
-
-    /// A `<switch>`.
-    Switch(SwitchField),
-
-    /// A `<fd>`.
-    Fd(FdField),
-
-    /// A `<list>` containing file descriptors.
-    FdList(FdListField),
-
-    /// A `<exprfield>`.
-    Expr(ExprField),
-
-    /// An invented length field for a list that has no length.
-    ///
-    /// Fields of this kind are synthesised while resolving.
-    VirtualLen(VirtualLenField),
-}
-
-impl FieldDef {
-    /// Get the name of the field, if it has a name.
-    ///
-    /// This returns the `name` attribute of the original field. `<pad>`s do not have a name. For a
-    pub fn name(&self) -> Option<&str> {
-        match self {
-            Self::Pad(_) => None,
-            Self::Normal(normal_field) => Some(&normal_field.name),
-            Self::List(list_field) => Some(&list_field.name),
-            Self::Switch(switch_field) => Some(&switch_field.name),
-            Self::Fd(fd_field) => Some(&fd_field.name),
-            Self::FdList(fd_list_field) => Some(&fd_list_field.name),
-            Self::Expr(expr_field) => Some(&expr_field.name),
-            Self::VirtualLen(virtual_len_field) => Some(&virtual_len_field.name),
-        }
-    }
-
-    /// Get the type of the field.
-    ///
-    /// * For `<field>`s and `<exprfield>`s, this returns the type of the field.
-    /// * For `<list>`s, this returns the element type of the list.
-    /// * For virtual length fields, this returns the type of the length, which is a `CARD32`.
-    pub fn value_type(&self) -> Option<&FieldValueType> {
-        match self {
-            Self::Pad(_) => None,
-            Self::Normal(normal_field) => Some(&normal_field.type_),
-            Self::List(list_field) => Some(&list_field.element_type),
-            Self::Switch(_) => None,
-            Self::Fd(_) => None,
-            Self::FdList(_) => None,
-            Self::Expr(expr_field) => Some(&expr_field.type_),
-            Self::VirtualLen(virtual_len_field) => Some(&virtual_len_field.type_),
-        }
-    }
-
-    /// Get the size in bytes of this field on the wire.
-    ///
-    /// This returns `None` for fields that have a variable size
-    pub fn size(&self) -> Option<u32> {
-        match self {
-            Self::Pad(pad_field) => match pad_field.kind {
-                PadKind::Align(_) => None,
-                PadKind::Bytes(pad_size) => Some(u32::from(pad_size)),
-            },
-            Self::Normal(normal_field) => normal_field.type_.size(),
-            Self::List(list_field) => {
-                let element_size = list_field.element_type.size()?;
-                let list_len = list_field.length()?;
-                Some(element_size * list_len)
-            }
-            Self::Switch(switch_field) => switch_field.size(),
-            // fds are not serialized, they are sent with fd passing
-            Self::Fd(_) => Some(0),
-            Self::FdList(_) => Some(0),
-            Self::Expr(expr_field) => expr_field.type_.size(),
-            // not serialized
-            Self::VirtualLen(_) => Some(0),
-        }
-    }
-}
-
-/// A `<pad>` field.
-#[derive(Clone, Debug)]
-pub struct PadField {
-    /// The kind and size of padding
-    pub kind: PadKind,
-
-    /// The value of the `serialize` attribute.
-    ///
-    /// If no attribute was present, the value is `false`.
-    pub serialize: bool,
-}
-
-/// A `<field>` field.
-#[derive(Debug)]
-pub struct NormalField {
-    /// The name of the field.
-    pub name: String,
-
-    /// The kind of values that the field can take.
-    pub type_: FieldValueType,
-}
-
-/// A `<list>` field.
-#[derive(Debug)]
-pub struct ListField {
-    /// The name of the field.
-    pub name: String,
-
-    /// The kind of values that the field can take.
-    pub element_type: FieldValueType,
-
-    /// The expression describing the length of the list, if present.
-    pub length_expr: Option<Expression>,
-}
-
-impl ListField {
-    /// Does the list have a fixed/constant length.
-    ///
-    /// This function is true if there is only one possible length for the list.
-    pub fn has_fixed_length(&self) -> bool {
-        self.length().is_some()
-    }
-
-    /// Get the length of the list if it is statically known.
-    pub fn length(&self) -> Option<u32> {
-        if let Some(ref length_expr) = self.length_expr {
-            match length_expr {
-                Expression::Value(v) => Some(u32::try_from(*v).unwrap()),
-                Expression::Bit(bit) => Some(1 << *bit),
-                _ => None,
-            }
-        } else {
-            None
-        }
-    }
-}
-
-/// A `<switch>` field.
-#[derive(Debug)]
-pub struct SwitchField {
-    /// The name of the field.
-    pub name: String,
-
-    /// The expression on which the `<switch>`... well, switches.
-    pub expr: Expression,
-
-    /// The value of this switch's `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// The computed alignment for this switch.
-    ///
-    /// This field is only set after resolving.
-    pub alignment: OnceCell<ComplexAlignment>,
-
-    /// The kind of switch.
-    ///
-    /// All `cases` of this switch have to conform to this kind: Either they are all `<case>`s or
-    /// all `<bitcase>`s.
-    pub kind: SwitchKind,
-
-    /// The `cases` that the switch can have.
-    pub cases: Vec<SwitchCase>,
-
-    /// The list of `<paramref>`s that appear in descendants of this switch.
-    ///
-    /// This list is empty before the switch was resolved.
-    pub external_params: RefCell<Vec<ExternalParam>>,
-}
-
-impl SwitchField {
-    /// Get the size in bytes of this switch on the wire.
-    ///
-    /// This returns `None` if the switch does not have a fixed size. Only switches of kind
-    /// `<bitcase>` can have a fixed size.
-    pub fn size(&self) -> Option<u32> {
-        if self.kind != SwitchKind::Case {
-            return None;
-        }
-
-        let mut size = None;
-        for case in self.cases.iter() {
-            let case_size = case.size()?;
-            if let Some(size) = size {
-                if size != case_size {
-                    return None;
-                }
-            } else {
-                size = Some(case_size);
-            }
-        }
-        size
-    }
-}
-
-/// The possible kinds that a `<switch>` can have.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum SwitchKind {
-    /// All children of the `<switch>` are `<bitcase>`s.
-    BitCase,
-
-    /// All children of the `<switch>` are `<case>`s.
-    Case,
-}
-
-/// A single `<case>` or `<bitcase>` of a switch.
-#[derive(Debug)]
-pub struct SwitchCase {
-    /// The value of the `name` attribute, if present.
-    pub name: Option<String>,
-
-    /// All the expressions that this case contains.
-    pub exprs: Vec<Expression>,
-
-    /// The value of the `<required_start_align>` child, if any.
-    pub required_start_align: Option<Alignment>,
-
-    /// Alignment information about this case.
-    ///
-    /// This information is not available before the case was resolved.
-    pub alignment: OnceCell<ComplexAlignment>,
-
-    /// The list of fields of this case.
-    ///
-    /// The fields appear in the order in which they are defined.
-    pub fields: RefCell<Vec<FieldDef>>,
-
-    /// The list of `<paramref>`s that appear in descendants of this case.
-    ///
-    /// This list is empty before the case was resolved.
-    pub external_params: RefCell<Vec<ExternalParam>>,
-}
-
-impl SwitchCase {
-    /// Get the size in bytes of this case on the wire.
-    ///
-    /// This returns `None` if this case does not have a constant size.
-    fn size(&self) -> Option<u32> {
-        self.fields
-            .borrow()
-            .iter()
-            .try_fold(0, |sz, field| Some(sz + field.size()?))
-    }
-}
-
-/// A `<fd>` field.
-#[derive(Debug)]
-pub struct FdField {
-    /// The value of the `name` attribute.
-    pub name: String,
-}
-
-/// A `<list>` field with `type="fd"`.
-#[derive(Debug)]
-pub struct FdListField {
-    /// The value of the `name` attribute.
-    pub name: String,
-
-    /// The length of the list
-    pub length_expr: Expression,
-}
-
-impl FdListField {
-    /// Get the length of the list if it is statically known.
-    pub fn length(&self) -> Option<u32> {
-        match self.length_expr {
-            Expression::Value(v) => Some(u32::try_from(v).unwrap()),
-            Expression::Bit(bit) => Some(1 << bit),
-            _ => None,
-        }
-    }
-}
-
-/// A `<exprfield>` field.
-#[derive(Debug)]
-pub struct ExprField {
-    /// The value of the `name` attribute.
-    pub name: String,
-
-    /// The type of the field.
-    pub type_: FieldValueType,
-
-    /// The expression describing how the field is computed.
-    pub expr: Expression,
-}
-
-/// An invented length field for a list that has no length.
-///
-/// There is currently exactly one `<exprfield>` tag in xcb-proto:
-/// ```text
-/// <request name="QueryTextExtents" opcode="48">
-///   <exprfield type="BOOL" name="odd_length">
-///     <op op="&amp;"><fieldref>string_len</fieldref><value>1</value></op>
-///   </exprfield>
-///   <field type="FONTABLE" name="font" />
-///   <list type="CHAR2B" name="string" />
-///   <reply>
-/// [...]
-/// ```
-/// The `<exprfield>` above refers to a field `string_len` that does not exist. This refers to the
-/// implicit length of the `string` list.
-///
-/// To make this work, a `VirtualLenField` is invented during resolving and inserted in the
-/// representation of the request above.
-#[derive(Debug)]
-pub struct VirtualLenField {
-    /// The name of the virtual length field.
-    pub name: String,
-
-    /// The type of the field.
-    // FIXME: This is always BuiltInType::Card32, so this field can be removed
-    pub type_: FieldValueType,
-
-    /// The name of the referenced list.
-    pub list_name: String,
-}
-
-/// The type of a field's value.
-#[derive(Debug)]
-pub struct FieldValueType {
-    /// The actual type of the field's value.
-    pub type_: NamedTypeRef,
-
-    /// Constraints describing the possible values that the field can have.
-    ///
-    /// This represents the `enum`, `altenum`, `mask`, and `altmask` attributes of the XML.
-    pub value_set: FieldValueSet,
-}
-
-impl FieldValueType {
-    /// Get the size in bytes of this field on the wire.
-    ///
-    /// This returns `None` if the struct does not have a fixed size.
-    ///
-    /// # Panics
-    ///
-    /// Panics if this value type was not yet resolved.
-    pub fn size(&self) -> Option<u32> {
-        self.type_.def.get().unwrap().size()
-    }
-}
-
-/// A possible value set constraining a field's value.
-#[derive(Debug)]
-pub enum FieldValueSet {
-    /// The field can take any possible value of its type.
-    None,
-
-    /// The field can only takes values from the referenced value.
-    Enum(NamedTypeRef),
-
-    /// The field can take any possible value of its type, but values of the referenced enum have a
-    /// special meaning.
-    AltEnum(NamedTypeRef),
-
-    /// The field is a bitmask of the values of the referenced enum.
-    Mask(NamedTypeRef),
-
-    /// The field can take any possible value of its type, but the referenced enum gives special
-    /// meaning to some bits.
-    AltMask(NamedTypeRef),
-}
-
-/// The kind of padding a `<pad>` can represent.
-#[derive(Clone, Debug)]
-pub enum PadKind {
-    /// A fixed size padding of the given size (`bytes` attribute).
-    Bytes(u16),
-
-    /// An alignment padding aligning to the given alignment ( `align` attribute).
-    Align(u16),
-}
-
-/// A reference to some type.
-#[derive(Clone, Debug)]
-pub enum TypeRef {
-    /// A built-in type is referenced.
-    BuiltIn(BuiltInType),
-
-    /// A struct is referenced.
-    Struct(Weak<StructDef>),
-
-    /// A union is referenced.
-    Union(Weak<UnionDef>),
-
-    /// An event structure is referenced.
-    EventStruct(Weak<EventStructDef>),
-
-    /// A X11 ID is referenced.
-    Xid(Weak<XidTypeDef>),
-
-    /// A union of X11 IDs is referenced.
-    XidUnion(Weak<XidUnionDef>),
-
-    /// An enumeration is referenced.
-    Enum(Weak<EnumDef>),
-
-    /// A type alias is referenced.
-    Alias(Weak<TypeAliasDef>),
-}
-
-impl TypeRef {
-    /// Check if two types are structurally identical.
-    ///
-    /// This does e.g. not resolve type aliases, but instead checks that both sides reference the
-    /// same type alias.
-    pub fn same_as(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::BuiltIn(builtin1), Self::BuiltIn(builtin2)) => builtin1 == builtin2,
-            (Self::Struct(def1), Self::Struct(def2)) => weak_ptr_eq(def1, def2),
-            (Self::Union(def1), Self::Union(def2)) => weak_ptr_eq(def1, def2),
-            (Self::EventStruct(def1), Self::EventStruct(def2)) => weak_ptr_eq(def1, def2),
-            (Self::Xid(def1), Self::Xid(def2)) => weak_ptr_eq(def1, def2),
-            (Self::XidUnion(def1), Self::XidUnion(def2)) => weak_ptr_eq(def1, def2),
-            (Self::Enum(def1), Self::Enum(def2)) => weak_ptr_eq(def1, def2),
-            (Self::Alias(def1), Self::Alias(def2)) => weak_ptr_eq(def1, def2),
-            _ => false,
-        }
-    }
-
-    /// Get the size in bytes of this case on the wire.
-    ///
-    /// This returns `None` if the type does not have a constant size.
-    pub fn size(&self) -> Option<u32> {
-        match self {
-            Self::BuiltIn(builtin_type) => Some(builtin_type.size()),
-            Self::Struct(struct_def) => {
-                let struct_def = struct_def.upgrade().unwrap();
-                struct_def.size()
-            }
-            Self::Union(union_def) => {
-                let union_def = union_def.upgrade().unwrap();
-                Some(union_def.size())
-            }
-            Self::EventStruct(event_struct_def) => {
-                let event_struct_def = event_struct_def.upgrade().unwrap();
-                for allowed in event_struct_def.alloweds.iter() {
-                    for event in allowed.resolved.borrow().iter() {
-                        let event_def = event.get_original_full_def();
-                        if event_def.xge {
-                            return None;
-                        }
-                    }
-                }
-
-                Some(32)
-            }
-            Self::Xid(_) => Some(4),
-            Self::XidUnion(_) => Some(4),
-            // never used directly
-            Self::Enum(_) => unreachable!(),
-            Self::Alias(type_alias_def) => {
-                let type_alias_def = type_alias_def.upgrade().unwrap();
-                type_alias_def.old_name.def.get().unwrap().size()
-            }
-        }
-    }
-
-    /// Get the original type.
-    ///
-    /// This function resolves type aliases by calling [`TypeAliasDef::get_original_type`].
-    pub fn get_original_type(&self) -> Self {
-        match self {
-            Self::Alias(type_alias_def) => {
-                let type_alias_def = type_alias_def.upgrade().unwrap();
-                type_alias_def.get_original_type()
-            }
-            _ => self.clone(),
-        }
-    }
-}
-
-/// A built-in type that is not explicitly defined.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BuiltInType {
-    /// An unsigned integer with 8 bits.
-    Card8,
-
-    /// An unsigned integer with 16 bits.
-    Card16,
-
-    /// An unsigned integer with 32 bits.
-    Card32,
-
-    /// An unsigned integer with 64 bits.
-    Card64,
-
-    /// A signed integer with 8 bits.
-    Int8,
-
-    /// A signed integer with 16 bits.
-    Int16,
-
-    /// A signed integer with 32 bits.
-    Int32,
-
-    /// A signed integer with 64 bits.
-    Int64,
-
-    /// A single byte.
-    Byte,
-
-    /// A boolean value
-    Bool,
-
-    /// A single character.
-    ///
-    /// Characters are represented as bytes. It best not to think about non-ASCII characters.
-    Char,
-
-    /// A single precision floating point value with 32 bits as defined by IEEE 754.
-    Float,
-
-    /// A double precision floating point value with 64 bits as defined by IEEE 754.
-    Double,
-
-    /// A value with an unspecified encoding.
-    ///
-    /// One example for this is the argument to xproto's `ChangeProperty` request. Depending on
-    /// other values in the request, the arguments have type `Card8`, `Card16`, or `Card32`.
-    ///
-    /// Values of this type can be treated as bytes. In particular, the size of a single value is
-    /// one and larger values (like in the `ChangeProperty` example above) are composed of multiple
-    /// instances of the `Void` type.
-    Void,
-}
-
-impl BuiltInType {
-    /// Get the size in bytes of this type on the wire.
-    pub fn size(self) -> u32 {
-        match self {
-            Self::Card8 => 1,
-            Self::Card16 => 2,
-            Self::Card32 => 4,
-            Self::Card64 => 8,
-            Self::Int8 => 1,
-            Self::Int16 => 2,
-            Self::Int32 => 4,
-            Self::Int64 => 8,
-            Self::Byte => 1,
-            Self::Bool => 1,
-            Self::Char => 1,
-            Self::Float => 4,
-            Self::Double => 8,
-            Self::Void => 1,
-        }
-    }
-}
-
-/// An expression for computing some value.
-///
-/// This is a recursive type describing a tree of the expression.
-#[derive(Debug)]
-pub enum Expression {
-    /// An operation with two sub-expressions.
-    BinaryOp(BinaryOpExpr),
-
-    /// An operation with one sub-expression.
-    UnaryOp(UnaryOpExpr),
-
-    /// A reference to a field of the type in which this expression is evaluated.
-    FieldRef(FieldRefExpr),
-
-    /// A reference to an outer parameter.
-    ///
-    /// Currently, there is only a single `<paramref>` in xcb-proto:
-    /// The reply to xinput's `GetDeviceMotionEvents` request has a `num_axes` field and an
-    /// `events` list. This list contains structures of type `DeviceTimeCoord`. This structure in
-    /// turn has a list. The length of this list is a `<paramref>` to `num_axes`.
-    ///
-    /// Thus, a `<paramref>` is a reference to a field that is not even present in the type where
-    /// the expression is originally, but in some context that is even farther away.
-    ParamRef(ParamRefExpr),
-
-    /// A reference to a value of an enum.
-    EnumRef(EnumRefExpr),
-
-    /// The value of this expression is the number of bits that are set in the value of the inner
-    /// expression.
-    PopCount(Box<Expression>),
-
-    /// The value of this expression is the sum of some other expression evaluated for each element
-    /// of some list.
-    SumOf(SumOfExpr),
-
-    /// A reference to the current list element inside of a `SumOf` expression.
-    ListElementRef,
-
-    /// A constant and fixed value.
-    Value(u32),
-
-    /// A constant and fixed value specified in the form `1 << bit`.
-    Bit(u8),
-}
-
-/// An expression with two sub-expressions.
-#[derive(Debug)]
-pub struct BinaryOpExpr {
-    /// The operator that is applied to the two sub-expressions.
-    pub operator: BinaryOperator,
-
-    /// The left hand side of the expression.
-    pub lhs: Box<Expression>,
-
-    /// The right hand side of the expression.
-    pub rhs: Box<Expression>,
-}
-
-/// The possible binary operators
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum BinaryOperator {
-    /// Addition
-    Add,
-
-    /// Subtraction
-    Sub,
-
-    /// Multiplication
-    Mul,
-
-    /// (Integer) division
-    Div,
-
-    /// Logical and
-    And,
-
-    /// A logical left shift
-    Shl,
-}
-
-/// An expression with one sub-expression
-#[derive(Debug)]
-pub struct UnaryOpExpr {
-    /// The operator that is applied to the sub-expression.
-    pub operator: UnaryOperator,
-
-    /// The sub-expression.
-    pub rhs: Box<Expression>,
-}
-
-/// The possible unary operators
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum UnaryOperator {
-    /// Bitwise negation
-    Not,
-}
-
-/// An expression referencing a field from the surrounding context.
-#[derive(Debug)]
-pub struct FieldRefExpr {
-    /// The name of the field that is referenced.
-    pub field_name: String,
-
-    /// After resolution, this contains information about the referenced field.
-    pub resolved: OnceCell<ResolvedFieldRef>,
-}
-
-/// Information about a field reference.
-#[derive(Debug)]
-pub struct ResolvedFieldRef {
-    /// The kind of reference to another field.
-    pub ref_kind: FieldRefKind,
-
-    /// The type of the referenced field.
-    pub field_type: TypeRef,
-}
-
-/// Possible kinds of references to fields inside of an expression.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum FieldRefKind {
-    /// The reference is to a field of the expression where the field is evaluated.
-    LocalField,
-
-    /// The reference is to an external parameter from a surrounding context.
-    ExtParam,
-
-    /// The reference is used inside of a [`SumOfExpr`].
-    SumOfRef,
-}
-
-/// An expression referencing a parameter from a surrounding context.
-#[derive(Debug)]
-pub struct ParamRefExpr {
-    /// The name of the referenced field.
-    pub field_name: String,
-
-    /// The type of the referenced field.
-    pub type_: NamedTypeRef,
-}
-
-/// An expression referencing a value from an enumeration.
-#[derive(Debug)]
-pub struct EnumRefExpr {
-    /// The enum that is referenced.
-    pub enum_: NamedTypeRef,
-
-    /// The name of the enum's variant that is used.
-    pub variant: String,
-}
-
-/// An expression representing the sum over some referenced list.
-#[derive(Debug)]
-pub struct SumOfExpr {
-    /// The name of the list that is used.
-    pub field_name: String,
-
-    /// After resolving, a reference to the list that is used.
-    pub resolved_field: OnceCell<ResolvedFieldRef>,
-
-    /// The expression that is applied to each element of the list.
-    ///
-    /// If no expression is provided, the elements are summed as-is.
-    // FIXME: Can the Option be replaced with `Expression::ListElementRef`?
-    pub operand: Option<Box<Expression>>,
-}
-
-/// Documentation of some type.
-#[derive(Clone, Debug)]
-pub struct Doc {
-    /// A brief, one-line description.
-    pub brief: Option<String>,
-
-    /// A verbose description.
-    pub description: Option<String>,
-
-    /// A usage example. This will most likely assume libxcb's C API.
-    pub example: Option<String>,
-
-    /// Documentation of the individual fields.
-    pub fields: Vec<FieldDoc>,
-
-    /// The possible X11 errors that can occur.
-    pub errors: Vec<ErrorDoc>,
-
-    /// A reference to a related type.
-    pub sees: Vec<SeeDoc>,
-}
-
-/// Documentation for a single type.
-#[derive(Clone, Debug)]
-pub struct FieldDoc {
-    /// The name of the field.
-    pub name: String,
-
-    /// A description of this field.
-    pub doc: Option<String>,
-}
-
-/// Documentation of a X11 error that can occur.
-#[derive(Clone, Debug)]
-pub struct ErrorDoc {
-    /// The name of the error.
-    pub type_: String,
-
-    /// A description of when this error occurs.
-    pub doc: Option<String>,
-}
-
-/// A reference to another type for further information.
-#[derive(Clone, Debug)]
-pub struct SeeDoc {
-    /// The kind of type that is referenced.
-    // TODO: Turn this into an enum?
-    pub type_: String,
-
-    /// The name of the thing that is referenced, for example the name of a request.
-    pub name: String,
-}
-
-// Replacement for `Weak::ptr_eq`, which requires
-// Rust 1.39.
-fn weak_ptr_eq<T>(weak1: &Weak<T>, weak2: &Weak<T>) -> bool {
-    Rc::ptr_eq(&weak1.upgrade().unwrap(), &weak2.upgrade().unwrap())
 }
